+++ mnemos/__init__.py
@@ -11,7 +11,7 @@ Quick start:
     results = m.search("vector storage")
 """
 
-__version__ = "10.26.2"
+__version__ = "10.27.0"
 
 from .core import Mnemos
 from .storage.base import MnemosStore, Memory